Add test for cancelling single hearing
