+++ UI.AutomationTests/Admin/Booking/EditBookingTest.cs
@@ -1,20 +1,14 @@
-using UI.PageModels.Pages.Admin.Booking;
-
 namespace UI.AutomationTests.Admin.Booking;
 
 [Category("Daily")]
-public class EditBookingTest : AdminWebUiTest
+public class EditBookingTest : HearingTest
 {
-    private string _hearingIdString;
-
     [Test]
     public void should_update_booking_schedule_and_change_judge()
     {
         var hearingScheduledDateAndTime = DateUtil.GetNow(EnvConfigSettings.RunOnSaucelabs).AddMinutes(60);
         var hearingDto = HearingTestData.CreateHearingDtoWithEndpoints(judgeUsername:"auto_aw.judge_02@hearings.reform.hmcts.net",scheduledDateTime:hearingScheduledDateAndTime);
         var bookingDetailsPage = BookHearingAndGoToDetailsPage(hearingDto);
-        TestContext.WriteLine(
-            $"Attempting to book a hearing with the case name: {hearingDto.CaseName} and case number: {hearingDto.CaseNumber}");
         
         var newTime = hearingDto.ScheduledDateTime.AddMinutes(30);
         var summaryPage = bookingDetailsPage.UpdateSchedule(newTime, hearingDto.DurationHour, hearingDto.DurationMinute);
@@ -33,21 +27,4 @@ public class EditBookingTest : AdminWebUiTest
         bookingDetailPage.ValidateDetailsPage(hearingDto);
         Assert.Pass();
     }
-    
-    private BookingDetailsPage BookHearingAndGoToDetailsPage(BookingDto bookingDto)
-    {
-        var driver = VhDriver.GetDriver();
-        driver.Navigate().GoToUrl(EnvConfigSettings.AdminUrl);
-        var loginPage = new AdminWebLoginPage(driver, EnvConfigSettings.DefaultElementWait);
-        var dashboardPage = loginPage.Login(AdminLoginUsername, EnvConfigSettings.UserPassword);
-
-        var createHearingPage = dashboardPage.GoToBookANewHearing();
-        var summaryPage = createHearingPage.BookAHearingJourney(bookingDto, FeatureToggles.UseV2Api());
-        var confirmationPage = summaryPage.ClickBookButton();
-        TestHearingIds.Add(confirmationPage.GetNewHearingId());
-        TestContext.WriteLine($"Hearing  ID: {_hearingIdString}");
-        var bookingDetailsPage = confirmationPage.ClickViewBookingLink();
-        return bookingDetailsPage;
-    }
-    
 }
